Semi-automatic code cleanup *Explicit <> *Collapse identical catch blocks *Prefer try-with-resource *Prefer lambdas and method references *Collections.sort to List.sort *Unnecessary casts

+++ community/shell/src/main/java/org/neo4j/shell/util/json/JSONObject.java
@@ -1234,7 +1234,7 @@ public class JSONObject
         try
         {
             Object o = opt( key );
-            return o instanceof Number ? ((Number) o).doubleValue() : new Double( (String) o ).doubleValue();
+            return o instanceof Number ? ((Number) o).doubleValue() : new Double( (String) o );
         }
         catch ( Exception e )
         {
